refactor: extract transaction validation
