+++ magicblock-aperture/src/requests/http/mod.rs
@@ -15,13 +15,13 @@ use magicblock_core::{
 use magicblock_metrics::metrics::{AccountFetchOrigin, ENSURE_ACCOUNTS_TIME};
 use prelude::JsonBody;
 use solana_account::{AccountSharedData, ReadableAccount};
-use solana_message::VersionedMessage;
 use solana_pubkey::Pubkey;
 use solana_transaction::{
     sanitized::SanitizedTransaction, versioned::VersionedTransaction,
 };
 use solana_transaction_status::UiTransactionEncoding;
 use tracing::*;
+use transaction_validation::validate_supported_transaction_shape;
 
 use super::RpcRequest;
 use crate::{
@@ -30,10 +30,6 @@ use crate::{
 
 pub(crate) type HandlerResult = RpcResult<Response<JsonBody>>;
 
-// Solana's builtin-program filters in compute-budget processing assume program
-// indices fit within a packet-bounded pubkey table (1232 / 32 = 38).
-const MAX_RUNTIME_PROGRAM_ID_INDEX_EXCLUSIVE: usize =
-    1232 / size_of::<Pubkey>();
 const SYSTEM_PROGRAM_ID: Pubkey =
     Pubkey::from_str_const("11111111111111111111111111111111");
 
@@ -302,30 +298,6 @@ impl HttpDispatcher {
     }
 }
 
-fn validate_supported_transaction_shape(
-    transaction: &VersionedTransaction,
-) -> RpcResult<()> {
-    if let VersionedMessage::V0(message) = &transaction.message {
-        if !message.address_table_lookups.is_empty() {
-            return Err(RpcError::transaction_verification(
-                "v0 transactions with address lookup tables are not supported",
-            ));
-        }
-    }
-
-    for instruction in transaction.message.instructions() {
-        let program_id_index = usize::from(instruction.program_id_index);
-        if program_id_index >= MAX_RUNTIME_PROGRAM_ID_INDEX_EXCLUSIVE {
-            return Err(RpcError::transaction_verification(format!(
-                "unsupported program id index {program_id_index}; max supported is {}",
-                MAX_RUNTIME_PROGRAM_ID_INDEX_EXCLUSIVE - 1
-            )));
-        }
-    }
-
-    Ok(())
-}
-
 /// A prelude module to provide common imports for all RPC handler modules.
 mod prelude {
     pub(super) use magicblock_core::{link::accounts::LockedAccount, Slot};
@@ -389,6 +361,7 @@ pub(crate) mod mocked;
 pub(crate) mod request_airdrop;
 pub(crate) mod send_transaction;
 pub(crate) mod simulate_transaction;
+mod transaction_validation;
 
 // Magic Router compatibility methods.
 pub(crate) mod get_delegation_status;
@@ -398,15 +371,9 @@ mod tests {
     use std::sync::{Arc, Mutex};
 
     use magicblock_config::config::ChainLinkConfig;
-    use magicblock_core::{
-        coordination_mode::{switch_to_primary_mode, switch_to_replica_mode},
-        link::blocks::BlockHash,
+    use magicblock_core::coordination_mode::{
+        switch_to_primary_mode, switch_to_replica_mode,
     };
-    use solana_message::{
-        compiled_instruction::CompiledInstruction, legacy::Message,
-        MessageHeader, VersionedMessage,
-    };
-    use solana_signature::Signature;
     use test_kit::ExecutionTestEnv;
 
     use super::*;
@@ -451,11 +418,6 @@ mod tests {
         }
     }
 
-    const SYSTEM_PROGRAM_ID: Pubkey =
-        Pubkey::from_str_const("11111111111111111111111111111111");
-    const COMPUTE_BUDGET_ID: Pubkey =
-        Pubkey::from_str_const("ComputeBudget111111111111111111111111111111");
-
     #[tokio::test]
     async fn require_primary_rpc_method_allows_send_transaction_in_primary_mode(
     ) {
@@ -490,69 +452,6 @@ mod tests {
             error.to_string().contains(
                 "sendTransaction is only available while validator is primary"
             ),
-            "unexpected error: {error}"
-        );
-    }
-
-    #[test]
-    fn accepts_program_id_index_within_runtime_limit() {
-        let transaction = VersionedTransaction {
-            signatures: vec![Signature::default()],
-            message: VersionedMessage::Legacy(Message {
-                header: MessageHeader {
-                    num_required_signatures: 1,
-                    num_readonly_signed_accounts: 0,
-                    num_readonly_unsigned_accounts: 37,
-                },
-                account_keys: {
-                    let mut keys = vec![SYSTEM_PROGRAM_ID];
-                    keys.extend(std::iter::repeat_n(SYSTEM_PROGRAM_ID, 36));
-                    keys.push(COMPUTE_BUDGET_ID);
-                    keys
-                },
-                recent_blockhash: BlockHash::new_unique(),
-                instructions: vec![CompiledInstruction {
-                    program_id_index: 37,
-                    accounts: vec![],
-                    data: vec![],
-                }],
-            }),
-        };
-
-        validate_supported_transaction_shape(&transaction).unwrap();
-    }
-
-    #[test]
-    fn rejects_program_id_index_outside_runtime_limit() {
-        let transaction = VersionedTransaction {
-            signatures: vec![Signature::default()],
-            message: VersionedMessage::Legacy(Message {
-                header: MessageHeader {
-                    num_required_signatures: 1,
-                    num_readonly_signed_accounts: 0,
-                    num_readonly_unsigned_accounts: 38,
-                },
-                account_keys: {
-                    let mut keys = vec![SYSTEM_PROGRAM_ID];
-                    keys.extend(std::iter::repeat_n(SYSTEM_PROGRAM_ID, 37));
-                    keys.push(COMPUTE_BUDGET_ID);
-                    keys
-                },
-                recent_blockhash: BlockHash::new_unique(),
-                instructions: vec![CompiledInstruction {
-                    program_id_index: 38,
-                    accounts: vec![],
-                    data: vec![],
-                }],
-            }),
-        };
-
-        let error =
-            validate_supported_transaction_shape(&transaction).unwrap_err();
-        assert!(
-            error
-                .to_string()
-                .contains("unsupported program id index 38"),
             "unexpected error: {error}"
         );
     }
